+++ src/index.ts
@@ -1,5 +1,6 @@
 export { default as OneBlinkForm } from './OneBlinkForm'
 export { default as OneBlinkAutoSaveForm } from './OneBlinkAutoSaveForm'
+export { default as OneBlinkReadOnlyForm } from './OneBlinkReadOnlyForm'
 export { default as useIsMounted } from './hooks/useIsMounted'
 export { default as useBooleanState } from './hooks/useBooleanState'
 export { default as useNullableState } from './hooks/useNullableState'